move scenes to folder and start on game systems

+++ src/scenes/scene_menu.cpp
@@ -1,7 +1,7 @@
 #include "scene_menu.hpp"
 #include "scene_play.hpp"
 #include "scene_options.hpp"
-#include "game_engine.hpp"
+#include "../game_engine.hpp"
 #include <exception>
 
 void Scene_Menu::init()
@@ -31,7 +31,7 @@ void Scene_Menu::sRender()
     m_menuText.setFillColor(sf::Color::White);
     m_menuText.setPosition(100, 100);
     m_game->window().draw(m_menuText);
-    for (int i = 0; i < m_menuStrings.size(); i++)
+    for (size_t i = 0; i < m_menuStrings.size(); i++)
     {
         m_menuText.setString(m_menuStrings[i]);
         m_menuText.setCharacterSize(16);
